requirements.txt
antlr4-python3-runtime==4.9.2
pytest==8.3.2              # Cho testing